Make the whole main area
Placed the logo with it's original size and padded it to the appropriate
position.
Made the search box and rounded it to look like the original, also
adding the box-shadow on hover.
Added some icons before and after the search box, and made them enter
the input using negative margins.
To style the buttons, I shamelessly used the dev tools on google page
and copied most characteristics.